+++ bricks/downloader/go_requests.py
@@ -10,10 +10,8 @@ import urllib.parse
 import warnings
 from typing import Union
 
-from curl_cffi.requests import Cookies
-from requests.cookies import RequestsCookieJar
-
 from bricks.downloader import genesis
+from bricks.lib.cookies import Cookies
 from bricks.lib.request import Request
 from bricks.lib.response import Response
 from bricks.utils import pandora
@@ -77,7 +75,7 @@ class Downloader(genesis.Downloader):
                     Response(
                         content=response.content,
                         headers=response.headers,
-                        cookies=self.make_cookies(response.cookies),
+                        cookies=Cookies.by_jar(response.cookies),
                         url=response.url,
                         status_code=response.status_code,
                         request=Request(
@@ -92,19 +90,12 @@ class Downloader(genesis.Downloader):
             else:
                 res.content = response.content
                 res.headers = response.headers
-                res.cookies = self.make_cookies(response.cookies)
+                res.cookies = Cookies.by_jar(response.cookies)
                 res.url = response.url
                 res.status_code = response.status_code
                 res.request = request
 
                 return res
-
-    @staticmethod
-    def make_cookies(raw: RequestsCookieJar):
-        cookies = Cookies()
-        for cookie in raw:
-            cookies.set(name=cookie.name, value=cookie.value, domain=cookie.domain, path=cookie.path)
-        return cookies
 
 
 if __name__ == '__main__':
