+++ bricks/lib/queues/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+# @Time    : 2023-12-11 13:14
+# @Author  : Kem
+# @Desc    :
+
+from bricks.lib.queues.genesis import TaskQueue, Item
+from bricks.lib.queues.smart import SmartQueue
+from bricks.lib.queues.redis_ import RedisQueue
+from bricks.lib.queues.local import LocalQueue
